Update account_models.py
The issue is in the __str__ method, specifically in the return statement of self.account_name, but there is no attribute named account_name in the model. To fix this problem, you can modify the __str__ method to use an existing attribute. In your case, it seems that name is the appropriate attribute.